% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust.bin.pair.R
\docType{package}
\name{clust.bin.pair}
\alias{clust.bin.pair}
\alias{clust.bin.pair-package}
\title{clust.bin.pair: A package for clustered binary matched pair data}
\usage{
clust.bin.pair(ak, bk, ck, dk, method = "yang")
}
\arguments{
\item{ak}{Vector containing counts per group of Success/Success results.}

\item{bk}{Vector containing counts per group of Success/Fail results.}

\item{ck}{Vector containing counts per group of Fail/Success results.}

\item{dk}{Vector containing counts per group of Fail/Fail results.}

\item{method}{A character string specifying the method to calculate the
statistic. Must be one of "yang" (default), "durkalski", "obuchowski", "eliasziw".}
}
\value{
A list with class "htest" containing the following components:
}
\description{
The clust.bin.pair package privides statistical tests, and several utilities
for dealing with clustered binary matched pair data.

An adjustment to mcnemar's test for maginal homogeneity based on the method
of moments
}
\section{Fields}{

\describe{
\item{\code{statistic}}{the value of the test statistic with a name describing it.}

\item{\code{p.value}}{the p-value for the test.}

\item{\code{method}}{the type of test applied.}

\item{\code{data.name}}{a character string giving the names of the data.}
}}
\examples{
tc <- data.frame(nested.to.contingency(thyroids$x.pet, thyroids$x.spect))
clust.bin.pair(tc$ak, tc$bk, tc$ck, tc$dk, method="obuchowski")

pc <- psychiatry[, c('ah', 'bh', 'ch', 'dh')]
clust.bin.pair(pc$ah, pc$bh, pc$ch, pc$dh, method="eliasziw")

}

