% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durkalski.R
\name{durkalski.test}
\alias{durkalski.test}
\title{An implementation of Durkalski et al. 2003}
\usage{
durkalski.test(ak, bk, ck, dk)
}
\arguments{
\item{ak}{Vector containing counts per group of Success/Success results.}

\item{bk}{Vector containing counts per group of Success/Fail results.}

\item{ck}{Vector containing counts per group of Fail/Success results.}

\item{dk}{Vector containing counts per group of Fail/Fail results.}
}
\value{
Chi-square statistic
}
\description{
An adjustment to mcnemar's test for maginal homogeneity based on the method
of moments
}
\examples{

durkalski.test(c(1,3,0,1,2), c(51,67,44,58,47), c(18,26,43,21,28), c(0,2,1,3,4))

pc <- psychiatry[, c('ah', 'bh', 'ch', 'dh')]
durkalski.test(pc$ah, pc$bh, pc$ch, pc$dh)

tc <- data.frame(nested.to.contingency(thyroids$x.pet, thyroids$x.spect))
durkalski.test(tc$ak, tc$bk, tc$ck, tc$dk)

oc <- paired.to.contingency(list(obfuscation$subject, obfuscation$atom),
                            obfuscation$control, obfuscation$treatment)
durkalski.test(oc$ak, oc$bk, oc$ck, oc$dk)

}

