% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliasziw.R
\name{eliasziw.test}
\alias{eliasziw.test}
\title{An implementation of Eliasziw and Donner 1991}
\usage{
eliasziw.test(ak, bk, ck, dk)
}
\arguments{
\item{ak}{Vector containing counts per group of Success/Success results.}

\item{bk}{Vector containing counts per group of Success/Fail results.}

\item{ck}{Vector containing counts per group of Fail/Success results.}

\item{dk}{Vector containing counts per group of Fail/Fail results.}
}
\value{
Chi-square statistic
}
\description{
An adjustment to mcnemar's test for maginal homogeneity based on the
intracluster correlation coe cient (ICC).
}
\examples{

eliasziw.test(c(1,3,0,1,2), c(20,47,39,30,28), c(18,45,23,21,29), c(0,2,1,3,4))

pc <- psychiatry[, c('ah', 'bh', 'ch', 'dh')]
eliasziw.test(pc$ah, pc$bh, pc$ch, pc$dh)

tc <- data.frame(nested.to.contingency(thyroids$x.pet, thyroids$x.spect))
eliasziw.test(tc$ak, tc$bk, tc$ck, tc$dk)

oc <- paired.to.contingency(list(obfuscation$subject, obfuscation$atom),
                            obfuscation$control, obfuscation$treatment)
eliasziw.test(oc$ak, oc$bk, oc$ck, oc$dk)

}

