% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{nested.to.contingency}
\alias{nested.to.contingency}
\title{Convert between nested results and the canonical contingency tables}
\usage{
nested.to.contingency(t1, t2)
}
\arguments{
\item{t1}{lists of pre-treatment measures}

\item{t2}{lists of post-treatment measures}
}
\value{
Contingency tables represented in the rows of a matrix
}
\description{
Sum all concordant and discordant pairs from each nested group into a
contingency table.
}
\examples{

nested.to.contingency(thyroids$x.pet, thyroids$x.spect)

}

