% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yang.R
\name{yang.test}
\alias{yang.test}
\title{An implementation of Yang et al. 2010}
\usage{
yang.test(ak, bk, ck, dk)
}
\arguments{
\item{ak}{Vector containing counts per group of Success/Success results.}

\item{bk}{Vector containing counts per group of Success/Fail results.}

\item{ck}{Vector containing counts per group of Fail/Success results.}

\item{dk}{Vector containing counts per group of Fail/Fail results.}
}
\value{
Chi-square statistic
}
\description{
An adjustment to mcnemar's test for maginal homogeneity, extended from
Obuchowski 1998.
}
\examples{

yang.test(c(3,13,2,1,2), c(16,63,15,15,79), c(7,11,12,3,9), c(12,32,3,23,3))

pc <- psychiatry[, c('ah', 'bh', 'ch', 'dh')]
yang.test(pc$ah, pc$bh, pc$ch, pc$dh)

tc <- data.frame(nested.to.contingency(thyroids$x.pet, thyroids$x.spect))
yang.test(tc$ak, tc$bk, tc$ck, tc$dk)

oc <- paired.to.contingency(list(obfuscation$subject, obfuscation$atom),
                            obfuscation$control, obfuscation$treatment)
yang.test(oc$ak, oc$bk, oc$ck, oc$dk)

}

