% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.bin.pair.R
\docType{package}
\name{clust.bin.pair}
\alias{clust.bin.pair}
\alias{clust.bin.pair-package}
\title{clust.bin.pair: A package for clustered binary matched pair data}
\description{
The clust.bin.pair package privides statistical tests, and several utilities
for dealing with clustered binary matched pair data.
}

