% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust.bin.pair.R
\docType{package}
\name{clust.bin.pair}
\alias{clust.bin.pair}
\alias{clust.bin.pair-package}
\title{clust.bin.pair: Statistical tests for clustered binary matched pair data}
\description{
The clust.bin.pair package privides statistical tests and utilities
for dealing with clustered binary matched pair data.
}

