% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust.bin.pair.R
\name{clust.bin.pair}
\alias{clust.bin.pair}
\title{Statistical test for clustered matched pair data}
\usage{
clust.bin.pair(ak, bk, ck, dk, method = "yang")
}
\arguments{
\item{ak}{vector containing counts per group of Success/Success results.}

\item{bk}{vector containing counts per group of Success/Failure results.}

\item{ck}{vector containing counts per group of Failure/Success results.}

\item{dk}{vector containing counts per group of Failure/Failure results.}

\item{method}{a character string specifying the method to calculate the
statistic. Must be one of "yang" (default), "durkalski", "obuchowski", "eliasziw".}
}
\value{
A list with class "htest" containing the following components:

\item{statistic}{the value of the test statistic.}
\item{p.value}{the p-value for the test.}
\item{method}{the type of test applied.}
\item{data.name}{a character string giving the names of the data.}
}
\description{
A single interface for several adjustements to the mcnemar test for marginal
homogeneity that correct for clustered data.
}
\examples{

tc <- data.frame(nested.to.contingency(thyroids$x.pet, thyroids$x.spect))
clust.bin.pair(tc$ak, tc$bk, tc$ck, tc$dk, method="obuchowski")

pc <- psychiatry[, c('ah', 'bh', 'ch', 'dh')]
clust.bin.pair(pc$ah, pc$bh, pc$ch, pc$dh, method="eliasziw")

}

