% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obuchowski.R
\name{obuchowski.test}
\alias{obuchowski.test}
\title{An implementation of Obuchowski 1998}
\usage{
obuchowski.test(ak, bk, ck, dk)
}
\arguments{
\item{ak}{Vector containing counts per group of Success/Success results.}

\item{bk}{Vector containing counts per group of Success/Fail results.}

\item{ck}{Vector containing counts per group of Fail/Success results.}

\item{dk}{Vector containing counts per group of Fail/Fail results.}
}
\value{
Chi-square statistic
}
\description{
An adjustment to mcnemar's test for maginal homogeneity.
}
\examples{

obuchowski.test(c(1,3,0,1,2), c(56,64,75,95,87), c(7,1,2,3,9), c(0,2,1,3,3))

pc <- psychiatry[, c('ah', 'bh', 'ch', 'dh')]
obuchowski.test(pc$ah, pc$bh, pc$ch, pc$dh)

tc <- data.frame(nested.to.contingency(thyroids$x.pet, thyroids$x.spect))
obuchowski.test(tc$ak, tc$bk, tc$ck, tc$dk)

oc <- paired.to.contingency(list(obfuscation$subject, obfuscation$atom),
                            obfuscation$control, obfuscation$treatment)
obuchowski.test(oc$ak, oc$bk, oc$ck, oc$dk)

}

